chore: abstract pagination and sorting components
