+++ apps/web/src/app/[locale]/challenge/_components/comments/index.tsx
@@ -2,13 +2,6 @@
 
 import type { CommentRoot } from '@repo/db/types';
 import { Button } from '@repo/ui/components/button';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@repo/ui/components/select';
 import { toast } from '@repo/ui/components/use-toast';
 import { ChevronDown, ChevronLeft, ChevronRight, MessageCircle } from '@repo/ui/icons';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
@@ -20,33 +13,9 @@ import { CommentSkeleton } from './comment-skeleton';
 import { addComment } from './comment.action';
 import { getPaginatedComments, type PreselectedCommentMetadata } from './getCommentRouteData';
 import NoComments from './nocomments';
-
-const sortKeys = [
-  {
-    label: 'Newest Comments',
-    value: 'newest',
-    key: 'createdAt',
-    order: 'desc',
-  },
-  {
-    label: 'Oldest Comments',
-    value: 'oldest',
-    key: 'createdAt',
-    order: 'asc',
-  },
-  {
-    label: 'Most Votes',
-    value: 'votes',
-    key: 'vote',
-    order: 'desc',
-  },
-  {
-    label: 'Most Replies',
-    value: 'replies',
-    key: 'replies',
-    order: 'desc',
-  },
-] as const;
+import { sortKeys } from '~/utils/sorting';
+import { SortSelect } from '../sort-select';
+import { Pagination } from '../pagination';
 
 interface Props {
   preselectedCommentMetadata?: PreselectedCommentMetadata;
@@ -163,27 +132,7 @@ export function Comments({ preselectedCommentMetadata, rootId, type, expanded = 
             />
           </div>
           {(data?.comments.length ?? 0) > 0 && (
-            <div className="flex items-center gap-2 px-3 py-2">
-              <Select
-                value={sortKey.value}
-                defaultValue="newest"
-                onValueChange={(value) => {
-                  setSortKey(sortKeys.find((sk) => sk.value === value) ?? sortKeys[0]);
-                  setPage(1);
-                }}
-              >
-                <SelectTrigger className="w-[180px]">
-                  <SelectValue placeholder="Sort Key" />
-                </SelectTrigger>
-                <SelectContent>
-                  {sortKeys.map((sortKey, index) => (
-                    <SelectItem key={index} value={sortKey.value}>
-                      {sortKey.label}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
+            <SortSelect sortKey={sortKey} setSortKey={setSortKey} setPage={setPage} />
           )}
           {status === 'loading' && <CommentSkeleton />}
           <div className="flex-1">
@@ -215,64 +164,5 @@ export function Comments({ preselectedCommentMetadata, rootId, type, expanded = 
         </div>
       </div>
     </div>
-  );
-}
-
-function Pagination({
-  currentPage,
-  totalPages,
-  onClick,
-}: {
-  totalPages: number;
-  currentPage: number;
-  onClick: (page: number) => void;
-}) {
-  const maxVisiblePages = 5;
-  const halfVisiblePages = Math.floor(maxVisiblePages / 2);
-
-  let startPage = currentPage - halfVisiblePages;
-  let endPage = currentPage + halfVisiblePages;
-
-  if (startPage <= 0) {
-    startPage = 1;
-    endPage = Math.min(totalPages, maxVisiblePages);
-  }
-
-  if (endPage > totalPages) {
-    endPage = totalPages;
-    startPage = Math.max(1, totalPages - maxVisiblePages + 1);
-  }
-
-  const pages = Array.from({ length: endPage - startPage + 1 }, (_, i) => startPage + i);
-
-  return (
-    <nav className="justify-space-between flex items-center gap-2">
-      <Button
-        disabled={currentPage === 1}
-        onClick={() => onClick(Math.max(0, currentPage - 1))}
-        variant="ghost"
-      >
-        <ChevronLeft />
-      </Button>
-      {pages.map((page) => (
-        <Button
-          className={clsx('border-border dark:border-ring border', {
-            'bg-border dark:bg-neutral-700': page === currentPage,
-          })}
-          key={`pagination-${page}`}
-          onClick={() => onClick(page)}
-          variant="ghost"
-        >
-          {page}
-        </Button>
-      ))}
-      <Button
-        disabled={currentPage === totalPages}
-        onClick={() => onClick(Math.min(totalPages, currentPage + 1))}
-        variant="ghost"
-      >
-        <ChevronRight />
-      </Button>
-    </nav>
   );
 }
